requirements.txt
Flask
python-dotenv
openpyxl
google-genai
requests
playwright
playwright-stealth
gunicorn